// recovered/igreja-web3/src/components/BalanceDisplay.d.ts
import { FC } from 'react';
interface BalanceDisplayProps {
    walletBalance: string;
    safeBalance: string;
    account: string | null;
    safeAddress: string | null;
}
declare const BalanceDisplay: FC<BalanceDisplayProps>;
export default BalanceDisplay;
//# sourceMappingURL=BalanceDisplay.d.ts.map